+++ examples/resources/harness_platform_role_assignments/import.sh
@@ -1,0 +1,2 @@
+# Import using roleassignments id
+terraform import harness_platform_role_assignments.example <role_assignments_id>